Fuzzy Rule
A fuzzy rule can be of the form
IF A is B THEN C is D, where A and C are variables and B and D are fuzzy sets. I've extended these rules to allow fuzzy binary operators and, or, xor and unary operator not (remember to use brackets for unary operators), while adding "are" as alias for "is" (for natural writing). Also, introduced the ELSE after THEN.